fix: add common props to transformer. restructure: schema_tools
